refactor: move parseRelativeDateRange to DateParser
